Check whether a permission exists or not on old versions of android...
otherwise we might unnecessarily deny a permission that we have because
it was too new for that API version. TO-DO: add an overrideable method
that says when the permission doesn't exist that by default calls
onGranted but can be overridden to do something else.

Also, remove bug comments since code fix was verified.

Lowered visibility of onResult method since it does not need to be
public